// src/app/model/patientData.model.ts
export interface patientData {
  id: string;
  title: string;
  name: string;
  emailId: string;
  contactNo: string;
  gender: string;
  martialStatus: string;
  dob: string;
  age: string;
  weight: string;
  height: string;
  bmi: string;
  bloodGroup: string;
  alcoholConsumption: string;
  smokingHabits: string;
  allergies: string;
  currentMedications: string;
  pastMedications: string;
}


// id: string;
// title?: string;
// name?: string;
// emailId?: string;
// contactNo?: string;
// gender?: string;
// martialStatus?: string;
// dob?: string;
// age?: string;
// weight?: string;
// height?: string;
// bmi?: string;
// bloodGroup?: string;
// alcoholConsumption?: string;
// smokingHabits?: string;
// allergies?: string;
// currentMedications?: string;
// pastMedications?: string;